pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "opendata"
version = "0.1.0"
description = "OpenData registry SDK and CLI"
readme = "opendata.md"
requires-python = ">=3.9"
dependencies = [
  "pandas>=2.1",
  "pyarrow>=14.0",
  "requests>=2.31",
  "PyYAML>=6.0",
]

[project.optional-dependencies]
dev = [
  "mypy>=1.8",
  "pytest>=7.0",
  "ruff>=0.4",
  "types-requests",
]
r2 = ["boto3>=1.34"]

[project.scripts]
od = "opendata.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.pytest.ini_options]
testpaths = ["tests"]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
check_untyped_defs = true
no_implicit_optional = true

[[tool.mypy.overrides]]
module = [
  "boto3.*",
  "botocore.*",
  "pandas.*",
  "pyarrow.*",
  "requests.*",
  "yaml.*",
]
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B"]
ignore = ["UP045"]

[tool.ruff.lint.isort]
known-first-party = ["opendata"]
